Removing print statement, adding CSS comment, renaming configs file
