+++ app.py
@@ -20,7 +20,7 @@ import dash
 import diskcache
 from dash import DiskcacheManager
 
-from app_configs import APP_TITLE, THEME_COLOR, THEME_COLOR_SECONDARY
+from demo_configs import APP_TITLE, THEME_COLOR, THEME_COLOR_SECONDARY
 from demo_interface import create_interface
 
 # Essential for initializing callbacks. Do not remove.
